% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riskmetric.R
\name{assessment}
\alias{assessment}
\title{Create a}
\usage{
assessment(assessment)
}
\arguments{
\item{assessment}{An object created by \code{pkg_assess()} (\code{list_of_pkg_metric})}
}
\value{
A data.frame with simplified objects but no scoring of the assessment.
}
\description{
Vectorize assessment to be able to add it to a report for a package or the PACKAGES file.
}
