% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_reporter}
\alias{check_reporter}
\title{Reports check results}
\usage{
check_reporter(path)
}
\arguments{
\item{path}{Path to check results folder (aka \code{pkgname.Rcheck}).}
}
\value{
A base report
}
\description{
Experimental function to use the default log to create a report.
}
\examples{
# Requires the package to be installed (or be in inst/ folder)
# check_reporter("checks/riskmetric.Rcheck")

}
\seealso{
\code{\link[tools:check_packages_in_dir]{tools::check_packages_in_dir()}}
}
