% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_utils.R
\name{create_info_card}
\alias{create_info_card}
\title{Create a info card for the report}
\usage{
create_info_card(title, header = NULL, text = NULL, extra_class = NULL)
}
\arguments{
\item{title}{text to include in the card title}

\item{header}{text or html content to be included in the card header}

\item{text}{content to be included as class card text}

\item{extra_class}{whether to add additional class for the card div in order to modify style}
}
\description{
Create a info card for the report
}
\details{
check https://getbootstrap.com/docs/5.0/components/card/ for class names and card content
To style the card, assign a custom class and style the card within the quarto theme
}
\keyword{internal}
