% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{environ_report}
\alias{environ_report}
\title{Export environmental variables}
\usage{
environ_report(show = NULL, exclude = NULL)
}
\arguments{
\item{show}{Character vector of environmental variables to show.}

\item{exclude}{Character vector of environmental variables to hide.}
}
\value{
A character vector with environmental variables.
}
\description{
Export environmental variables
}
\examples{
environ_report()
}
