% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_utils.R
\name{extract_license}
\alias{extract_license}
\title{analyze license output from risk assessment to display in cards}
\usage{
extract_license(assessment)
}
\arguments{
\item{assessment}{object returning from riskmetric assessment, it can also be data.frame from val.report::assessment}
}
\value{
list with main card license string and footer if available
}
\description{
analyze license output from risk assessment to display in cards
}
\details{
The function analyzes if the license is standard. If it is standard, then it will see if there is available
information for the footer. The footer information is only the multiple version number for the moment.
}
\keyword{internal}
