% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_utils.R
\name{get_image_name}
\alias{get_image_name}
\title{Return the Docker Image name to be printed on the report}
\usage{
get_image_name(params)
}
\description{
Return the Docker Image name to be printed on the report
}
\keyword{internal}
