% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_pkg_origin}
\alias{get_pkg_origin}
\title{Return string to display Origin depending on source of the assessment}
\usage{
get_pkg_origin(source)
}
\arguments{
\item{source}{return value of function \code{pkg_ref}}
}
\value{
depending on source transformed string for the origin
}
\description{
Return string to display Origin depending on source of the assessment
}
\details{
Please check \code{riskmetric::pkg_ref_class_hierarchy} and \code{riskmetric::pkg_ref} code
to see the allowed values and structure
}
\keyword{internal}
