% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riskmetric.R
\name{is_risk_error}
\alias{is_risk_error}
\title{Checks if it is a riskmetric error or not.}
\usage{
is_risk_error(x)
}
\arguments{
\item{x}{An object}
}
\value{
A logical value: if it is not a riskmetric metric it returns \code{NA}, but normally \code{TRUE} or \code{FALSE}.
}
\description{
Provide information whether the object metric is an error or not.
}
\examples{
is_risk_error(1)
}
