% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-knitr.R
\name{knitr_logger}
\alias{knitr_logger}
\title{Create a knitr log function}
\usage{
knitr_logger()
}
\value{
A \code{function} accepting \code{...} arguments, which will be used for
printing out to the console while rendering the knitr document. Character
values are logged directly, while any other object is printed as though
being printed to the console. To render character values as they would
be printed, mark them as "AsIs" using \code{\link[=I]{I()}}.
}
\description{
Sets necessary knitr hooks and returns a logging function that will emit
messages to the console during knitting.
}
\note{
\code{val.report} \code{knitr} utilities are exported for use in reports maintained by the R Validation Hub. If you choose to use these functions for other purposes, be aware that these are not considered stable for broader use.
}
\examples{
knitr_logger()
}
