% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-knitr.R
\name{knitr_logger_styles}
\alias{knitr_logger_styles}
\title{Various elements used for tailoring output style to command-line utility}
\usage{
knitr_logger_styles
}
\description{
Various elements used for tailoring output style to command-line utility
}
