% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_knitr.R
\name{knitr_mutable_header}
\alias{knitr_mutable_header}
\title{Create mutable header object}
\usage{
knitr_mutable_header()
}
\arguments{
\item{params}{Optionally, provide default parameters to initialize with}

\item{envir}{Only used when \code{params} is not provided, as the source for where
to try to discover the default knitr frontmatter parameters.}
}
\description{
Inject a custom document rendering hook into the knitr engine and return
a mutable environment that we can modify, which will be used to update the
knitr document header upon render completion.
}
\note{
\code{val.report} \code{knitr} utilities are exported for use in reports maintained by the R Validation Hub. If you choose to use these functions for other purposes, be aware that these are not considered stable for broader use.
}
