% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_utils.R
\name{knitr_update_options}
\alias{knitr_update_options}
\title{Handler for complex option passing through from a quarto parameter}
\usage{
knitr_update_options(opts)
}
\description{
Importantly, handles S7 objects which cannot be passed through as a quarto
parameter because they can not be deparsed. Allows passing arbitrary
expressions using the \code{!expr} prefix used by \code{yaml}.
}
