% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{options_report}
\alias{options_report}
\title{Export options}
\usage{
options_report(show = NULL, exclude = NULL)
}
\arguments{
\item{show}{Character vector of options to show.}

\item{exclude}{Character vector of options variables to hide.}
}
\value{
A list with the options
}
\description{
Export options
}
\examples{
options_report()
}
