% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_table.R
\name{prepare_namespace_table}
\alias{prepare_namespace_table}
\title{Prepare subtable for metric based risk assessment main table}
\usage{
prepare_namespace_table(riskmetric_assessment)
}
\arguments{
\item{riskmetric_assessment}{object coming from rismetric::pkg_assess function}
}
\description{
Prepare subtable for metric based risk assessment main table
}
\keyword{internal}
