% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riskmetric.R
\name{replace_zero_or_false_by_no}
\alias{replace_zero_or_false_by_no}
\title{Replace 0 or FALSE by no in the summary table}
\usage{
replace_zero_or_false_by_no(value)
}
\description{
Replace 0 or FALSE by no in the summary table
}
\keyword{internal}
