% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session.R
\name{session}
\alias{session}
\title{Collect session information}
\usage{
session()
}
\value{
A list with time, r_version, session_info, platform, machine,
rng_kind, capabilities, external_software and graphics_software information.
}
\description{
Catalogs characteristics of the R session and execution environment.
}
\examples{
s <- session()
str(s, 2)
}
