% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riskmetric.R
\name{summary_table}
\alias{summary_table}
\title{Summary table}
\usage{
summary_table(risk)
}
\arguments{
\item{risk}{The output of \code{assessment()}.}
}
\value{
A data.frame with the two columns one for the fields and one for the values.
}
\description{
Summary table
}
