% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/val-report-package.R
\docType{package}
\name{val.report-package}
\alias{val.report}
\alias{val.report-package}
\title{val.report: Validation Reports}
\description{
\if{html}{\figure{logo.svg}{options: style='float: right' alt='logo' width='120'}}

Documents package checks for validation in a pharmaceutic context.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pharmaR/val.report}
  \item \url{https://pharmar.github.io/val.report/}
  \item Report bugs at \url{https://github.com/pharmaR/val.report/issues}
}

}
\author{
\strong{Maintainer}: Lluís Revilla Sancho \email{lluis.revilla_sancho@roche.com} (\href{https://orcid.org/0000-0001-9747-2570}{ORCID})

Authors:
\itemize{
  \item Lluís Revilla Sancho \email{lluis.revilla_sancho@roche.com} (\href{https://orcid.org/0000-0001-9747-2570}{ORCID})
}

Other contributors:
\itemize{
  \item R Validation Hub \email{r-val-hub-maintainers@lists.r-consortium.org} [copyright holder]
}

}
\keyword{internal}
