% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_knitr.R
\name{knit_print.knitr_log}
\alias{knit_print.knitr_log}
\title{Special handler for emitting knitr logs}
\usage{
knit_print.knitr_log(x, ...)
}
\arguments{
\item{x}{An R object to be printed}

\item{...}{Additional arguments passed to the S3 method. Currently ignored,
except two optional arguments \code{options} and \code{inline}; see
the references below.}
}
\description{
Special handler for emitting knitr logs
}
