% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_knitr.R
\name{knitr_logger}
\alias{knitr_logger}
\title{Create a knitr log function}
\usage{
knitr_logger()
}
\value{
A \code{function} accepting \code{...} arguments, which will be used for
printing out to the console while rendering the knitr document. Character
values are logged directrly, while any other object is printed as though
being printed to the console. To render character values as they would
be printed, mark them as "AsIs" using \code{\link[=I]{I()}}.
}
\description{
Sets necessary knitr hooks and returns a logging function that will emit
messages to the console during knitting.
}
