% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_utils.R
\name{knitr_logger}
\alias{knitr_logger}
\title{Create a knitr log function}
\usage{
knitr_logger()
}
\description{
Sets necessary knitr hooks and returns a logging function that will emit
messages to the console during knitting.
}
