% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_knitr.R
\docType{data}
\name{knitr_logger_styles}
\alias{knitr_logger_styles}
\title{Various elements used for tailoring output style to command-line utility}
\format{
An object of class \code{list} of length 2.
}
\usage{
knitr_logger_styles
}
\description{
Various elements used for tailoring output style to command-line utility
}
\keyword{datasets}
