% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_utils.R
\name{knitr_mutable_header}
\alias{knitr_mutable_header}
\title{Create mutable header object}
\usage{
knitr_mutable_header()
}
\description{
Inject a custom document rendering hook into the knitr engine and return
a mutable environment that we can modify, which will be used to update the
knitr document header upon render completion.
}
