% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_knitr.R
\name{knitr_update_options}
\alias{knitr_update_options}
\title{Handler for complex option passing through from a quarto parameter}
\usage{
knitr_update_options(opts)
}
\arguments{
\item{opts}{A \code{list} of options. For any complex values that cannot be passed
through the \code{yaml} frontmatter, you may pass them as an expression string
such as \code{"!expr mtcars"}.}
}
\value{
The \code{opts} list after parsing any complex expressions. This function
is used primarily for modifying the global state by calling \code{\link[=options]{options()}}
}
\description{
Importantly, handles S7 objects which cannot be passed through as a quarto
parameter because they can not be deparsed. Allows passing arbitrary
expressions using the \code{!expr} prefix used by \code{yaml}.
}
