% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporter.R
\name{package_report}
\alias{package_report}
\title{Generates an R Validation Hub package report}
\usage{
package_report(
  package = NULL,
  session = NULL,
  options = val_options(plus = c("width")),
  template = system.file("report", "template.qmd", package = "val.report"),
  output_dir = getwd(),
  ...
)
}
\description{
Generates an R Validation Hub package report
}
\examples{
options(
  val.meter.logs = TRUE,
  val.meter.policy = val.meter::policy(permissions = TRUE)
)

package_report(package = "path/to/package")

}
